StarRating: Still save rating if scroll change is less than 10px in either direction
This fixes firefox mobile.

Chromium on mobile stops the page from scrolling with a grace area while dragging the stars, but firefox doesn't do this, which breaks the (if scrolled dont save logic), so added the 10px grace into logic.

Future fix could be to stop the scrolling to mimic chromium behaviour for firefox mobile too, but that sound hard. Maybe firefox can fix this.